requirements.txt
accelerate==0.26.0
bert-score==0.3.13
faiss-cpu==1.7.4
huggingface-hub==0.20.3
llama-cpp-python>=0.3.13
nltk==3.8.1
numpy==1.26.4
rouge-score==0.1.2
scikit-learn==1.4.2
scipy==1.11.4
sentence-transformers==2.7.0
sentencepiece==0.1.99
tqdm==4.66.1
transformers==4.36.2
